requirements.txt
Flask==3.1.1
gunicorn==21.2.0
geopy==2.4.1
timezonefinder==6.2.0
pytz==2025.2
pyswisseph==2.10.3.2
